+++ test/cinder_ui/hooks_test.exs
@@ -14,7 +14,7 @@ defmodule CinderUI.HooksTest do
   test "app_js_snippet returns hook integration snippet" do
     snippet = Hooks.app_js_snippet()
 
-    assert snippet =~ "import { CinderUIHooks } from \"./cinder_ui\""
+    assert snippet =~ "import { CinderUIHooks } from \"cinder_ui\""
     assert snippet =~ "Object.assign(Hooks, CinderUIHooks)"
   end